+++ src/main/java/blackjack/domain/Dealer.java
@@ -1,5 +1,13 @@
 package blackjack.domain;
 
-public class Dealer {
+public class Dealer extends Gamer {
+    private static final String name = "딜러";
 
+    public Dealer() {
+        super(name);
+    }
+
+    public Card getFirstCard() {
+        return cards.getValues().get(0);
+    }
 }
